+++ src/base/style.c
@@ -174,25 +174,27 @@ static uint32_t to_border(const char* value) {
 }
 
 static uint32_t to_icon_at(const char* value) {
-  uint32_t icon_at = ICON_AT_AUTO;
-
   if (strstr(value, "cent")) {
-    icon_at = ICON_AT_CENTRE;
-  }
-  if (strstr(value, "left")) {
-    icon_at = ICON_AT_LEFT;
-  }
-  if (strstr(value, "right")) {
-    icon_at = ICON_AT_RIGHT;
-  }
-  if (strstr(value, "top")) {
-    icon_at = ICON_AT_TOP;
-  }
-  if (strstr(value, "bottom")) {
-    icon_at = ICON_AT_BOTTOM;
+    return ICON_AT_CENTRE;
+  } else if (strstr(value, "left_top")) {
+    return ICON_AT_LEFT_TOP;
+  } else if (strstr(value, "left_bottom")) {
+    return ICON_AT_LEFT_BOTTOM;
+  } else if (strstr(value, "right_top")) {
+    return ICON_AT_RIGHT_TOP;
+  } else if (strstr(value, "right_bottom")) {
+    return ICON_AT_RIGHT_BOTTOM;
+  } else if (strstr(value, "left")) {
+    return ICON_AT_LEFT;
+  } else if (strstr(value, "right")) {
+    return ICON_AT_RIGHT;
+  } else if (strstr(value, "top")) {
+    return ICON_AT_TOP;
+  } else  if (strstr(value, "bottom")) {
+    return ICON_AT_BOTTOM;
   }
 
-  return icon_at;
+  return ICON_AT_LEFT; 
 }
 
 ret_t style_normalize_value(const char* name, const char* value, value_t* out) {